requirements.txt
six==1.15.0
lxml==4.5.2
requests==2.24.0
pandas==1.1.2
coloredlogs==14.0
arxivcheck==0.3.2
beautifulsoup4==4.9.3
bibtexparser==1.2.0
Pillow==8.0.1
PyYAML==5.3.1
selenium==3.141.0
title2bib==0.4.1
